refactor: Rename create_mock_hassette to create_hassette_stub (#259)
Rename mock_hassette.py → web_mocks.py and create_mock_hassette() →
create_hassette_stub() to clarify the web/API MagicMock stub role vs
HassetteHarness. Add "Choosing a Mock Strategy" guide to TESTING.md
and a Test Infrastructure section to CLAUDE.md.
